+++ osu.Native/Structures/Difficulty/NativeTaikoDifficultyAttributes.cs
@@ -5,7 +5,7 @@ namespace osu.Native.Structures.Difficulty;
 internal struct NativeTaikoDifficultyAttributes(TaikoDifficultyAttributes attributes)
 {
   public double StarRating = attributes.StarRating;
-  public double MaxCombo = attributes.MaxCombo;
+  public int MaxCombo = attributes.MaxCombo;
   public double RhythmDifficulty = attributes.RhythmDifficulty;
   public double ReadingDifficulty = attributes.ReadingDifficulty;
   public double ColourDifficulty = attributes.ColourDifficulty;
